+++ util/tracing/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `MaybeTracer` type to more easily support optional tracers.
 
 ## 0.3.1 - 2019-07-15
 ### Added